added a status check in the orders

+++ src/main/java/az/ingress/hrms/dto/orderPersonBonus/CreateOrderPersonBonusRequest.java
@@ -27,7 +27,7 @@ public class CreateOrderPersonBonusRequest {
 
     @NotNull(message = "Bonus ID cannot be null.")
     @Positive(message = "Bonus ID must be positive.")
-    private Long bonusTypeId;
+    private Integer bonusTypeId;
 
     @NotNull
     private BonusCalculationType calculationType;
